pyproject.toml
[project]
name = "wfg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "openai>=1.88.0",
    "pocketflow>=0.0.2",
    "tiktoken>=0.9.0",
    "tree-sitter>=0.24.0",
    "networkx>=3.4.2",
    "tqdm>=4.67.1",
    "python-dotenv>=1.0.0",
]
